requirements/test.txt
-r base.txt

# Testing
# ------------------------------------------------------------------------------
mypy
pytest
pytest-sugar

# Code quality
# ------------------------------------------------------------------------------
flake8
coverage
black
pylint-django
pylint-celery

# Django
# ------------------------------------------------------------------------------
factory-boy

django-extensions
django-coverage-plugin
django-test-plus
pytest-django
